added dir with files the standard way
